requirement.txt
pywebview
flask
paramiko
pymysql
pymssql
cx_oracle
pycryptodome
pywin32
pyautogui
pyinstaller
requests


torch
pydub
ffprobe
pyinstaller
ipython
torchaudio
flask
soundfile
numpy
ffmpeg